pyproject.toml
[project]
name = "human-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "markitdown[pdf]>=0.1.4",
    "openai-whisper>=20250625",
    "scipy>=1.17.0",
    "torch>=2.9.1",
    "transformers>=4.57.5",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.30.6"
]
